backend/requirements.txt
# Archon Backend

# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Celery & Redis
celery[redis]==5.3.6
redis==5.0.1
flower==2.0.1

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
psycopg2-binary==2.9.9

# Search Engines
meilisearch==0.28.4
qdrant-client==1.7.0

# AI & Embeddings
google-generativeai==0.8.0
tiktoken==0.5.2

# Document Processing
pymupdf==1.23.18
pillow==10.2.0
pytesseract==0.3.10
python-magic==0.4.27

# Archive Extraction
rarfile==4.1
py7zr==0.20.8

# Email Parsing
eml-parser==1.17.5
compressed-rtf==1.0.6

# NLP / Entity Extraction
spacy==3.7.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Database Migrations
alembic==1.13.1

# Testing
pytest==8.2.0
pytest-asyncio==0.24.0
